+++ constants/user-constants.js
@@ -1,0 +1,5 @@
+export const emailRegexp = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
+
+export default {
+  emailRegexp,
+};
