+++ controllers/auth-controller/index.js
@@ -1,16 +1,16 @@
 import { getCurrentUser } from "./getCurrentUser.js";
-import { singIn } from "./singIn.js";
+import { signIn } from "./signIn.js";
 import { logout } from "./logout.js";
-import { singUp } from "./singUp.js";
+import { signUp } from "./signUp.js";
 import { updateUser } from "./updateUser.js";
 
 import { ctrlWrapper } from "../../helpers/index.js";
 
 const authController = {
   getCurrentUser: ctrlWrapper(getCurrentUser),
-  singIn: ctrlWrapper(singIn),
+  signIn: ctrlWrapper(signIn),
   logout: ctrlWrapper(logout),
-  singUp: ctrlWrapper(singUp),
+  signUp: ctrlWrapper(signUp),
   updateUser: ctrlWrapper(updateUser),
 };
