+++ controllers/auth-controller/signIn.js
@@ -1,0 +1,3 @@
+export const signIn = async (req, res) => {};
+
+export default { signIn };
