--- controllers/auth-controller/singIn.js
@@ -1,3 +1,0 @@
-export const singIn = async (req, res) => {};
-
-export default { login };
