+++ routes/api/auth-router.js
@@ -3,14 +3,14 @@ import { authController } from "../../controllers/auth-controller/index.js";
 
 const router = express.Router();
 
-router.post("/register", authController.register);
+router.post("/signup", authController.signUp);
 
-router.post("/login", authController.singUp);
+// router.post("/login", authController.signIn);
 
-router.get("/currentUser", authController.getCurrentUser);
+// router.get("/currentUser", authController.getCurrentUser);
 
-router.post("/logout", authController.logout);
+// router.post("/logout", authController.logout);
 
-router.patch("/", authController.updateUser);
+// router.patch("/", authController.updateUser);
 
 export { router };
